fix: implement login function in MemberListViewModel
